cleanup preview routes
